refactor: organize main index
