+++ main/global-shortcuts.js
