+++ main/index.js
@@ -1,8 +1,4 @@
 var isDev = require('electron-is-dev')
-var sentry
-if (!isDev) {
-  sentry = require('../lib/sentry')
-}
 var { app, ipcMain } = require('electron')
 var Config = require('electron-store')
 var get = require('lodash.get')
@@ -10,21 +6,32 @@ var xtend = require('xtend')
 var userConfig = require('./config')
 var menu = require('./menu')
 var artwork = require('./artwork')
-var GlobalShortcuts = require('./globalShortcuts')
+var GlobalShortcuts = require('./global-shortcuts')
 var makeTrackDict = require('./track-dict')
 var audio = require('./windows/audio')
 var player = require('./windows/player')
 var AudioLibrary = require('./lib/audio-library')
 var ipcLogger = require('electron-ipc-log')
-var globalShortcuts = new GlobalShortcuts()
 var autoUpdater = require('electron-updater').autoUpdater
 var log = require('electron-log')
 
+// handle uncaught exceptions before calling any functions
+process.on('uncaughtException', (err) => {
+  log.error(err)
+})
+
+// register sentry if in production
+var sentry = isDev ? null : require('../lib/sentry')
+
+// log IPC events
 ipcLogger(event => {
   var { channel, data } = event
   if (channel === 'timeupdate') return
   log.info('✨  ipc', channel, ...data)
 })
+
+var globalShortcuts = new GlobalShortcuts()
+var windows = [player, audio]
 
 var persist = new Config({ name: 'hyperamp-persist' })
 var libraryPersist = new Config({ name: 'hyperamp-library' })
@@ -63,187 +70,32 @@ app.on('ready', function appReady () {
   audio.init()
   player.init()
   artwork.init()
+
   globalShortcuts.init({
     MediaNextTrack: next,
     MediaPreviousTrack: prev,
     MediaPlayPause: playPause
   })
 
-  // Emit things to all windows
-  var windows = [player, audio]
-  function broadcast (/* args */) {
-    var args = [].slice.call(arguments, 0)
-    windows.forEach((winObj) => {
-      if (winObj.win) winObj.win.send.apply(winObj.win, args)
-    })
-  }
-
-  function volume (ev, level) {
-    // player -> audio
-    state.volume = level
-    if (audio.win) audio.win.send('volume', level)
-  }
-
+  // register IPC handlers
   ipcMain.on('volume', volume)
-
-  function queue (ev, newIndex) {
-    var newTrack = al.queue(newIndex)
-    log.info(newTrack)
-    broadcast('new-track', newTrack)
-    if (player.win) {
-      player.win.send('new-index', al.index)
-      player.win.send('is-new-query', al.isNewQuery)
-    }
-    if (state.playing) { broadcast('play') }
-    updateArtwork()
-  }
-
-  function updateArtwork () {
-    if (!get(al, `currentTrack.artwork`)) {
-      artwork.cache.getPath(al.currentKey, handleGetPath)
-    }
-  }
-
-  function handleGetPath (err, blobPath) {
-    if (err) return log.error(err)
-    al.currentTrack.artwork = blobPath
-    if (player.win) {
-      player.win.send('new-track', al.currentTrack)
-      player.win.send('new-index', al.index)
-    }
-  }
-
   ipcMain.on('queue', queue)
-
-  function play () {
-    state.playing = true
-    broadcast('play')
-  }
-
   ipcMain.on('play', play)
-
-  function pause () {
-    state.playing = false
-    broadcast('pause')
-  }
-
   ipcMain.on('pause', pause)
-
-  function playPause () {
-    state.playing ? pause() : play()
-  }
-
-  function prev () {
-    broadcast('new-track', al.prev())
-    if (player.win) player.win.send('new-index', al.index)
-    if (state.playing) { broadcast('play') }
-    updateArtwork()
-  }
-
   ipcMain.on('prev', prev)
-
-  function next () {
-    broadcast('new-track', al.next())
-    if (player.win) player.win.send('new-index', al.index)
-    if (state.playing) { broadcast('play') }
-    updateArtwork()
-  }
-
   ipcMain.on('next', next)
-
-  function mute (ev) {
-    state.muted = true
-    broadcast('mute')
-  }
-
   ipcMain.on('mute', mute)
-
-  function unmute (ev) {
-    state.muted = false
-    broadcast('unmute')
-  }
-
   ipcMain.on('unmute', unmute)
-
-  function timeupdate (ev, currentTime) {
-    // audio -> player
-    state.currentTime = currentTime
-    if (player.win) player.win.send('timeupdate', currentTime)
-  }
-
   ipcMain.on('shuffle', shuffle)
-
-  function shuffle (ev) {
-    al.shuffle()
-    if (player.win) player.win.send('shuffle')
-  }
-
   ipcMain.on('unshuffle', unshuffle)
-
-  function unshuffle (ev) {
-    al.unshuffle()
-    if (player.win) player.win.send('unshuffle')
-  }
-
   ipcMain.on('timeupdate', timeupdate)
-
-  function seek (ev, newTime) {
-    // player -> audio
-    state.currentTime = newTime
-    if (audio.win) audio.win.send('seek', newTime)
-  }
-
   ipcMain.on('seek', seek)
-
-  function handleNewTracks (err, newTrackDict) {
-    state.loading = false
-    broadcast('loading', false)
-    if (err) return log.warn(err)
-    var newState = al.load(newTrackDict)
-    if (player.win) player.win.send('track-dict', newState.trackDict, newState.order, state.paths)
-    console.timeEnd('update-library')
-    log.info('Done scanning. Found ' + Object.keys(newState.trackDict).length + ' tracks.')
-  }
-
-  ipcMain.on('update-library', function (ev, paths) {
-    if (state.loading) state.loading.destroy()
-    log.info('Updating library with new path(s): ' + paths)
-    console.time('update-library')
-    state.paths = paths
-    state.loading = makeTrackDict(paths, handleNewTracks)
-    broadcast('loading', true)
-  })
-
-  function search (ev, searchString) {
-    if (player.win) {
-      player.win.send('track-order', al.search(searchString))
-      player.win.send('is-new-query', al.isNewQuery)
-    }
-  }
-
+  ipcMain.on('update-library', updateLibrary)
   ipcMain.on('search', search)
-
-  function recall () {
-    if (player.win) player.win.send('recall', al.recall(), al.searchTerm)
-  }
-
   ipcMain.on('recall', recall)
+  ipcMain.on('sync-state', syncState)
 
-  ipcMain.on('sync-state', (ev) => {
-    ev.sender.send('sync-state', {
-      trackDict: al.trackDict,
-      order: al.order,
-      paths: state.paths
-    })
-  })
-
-  if (process.platform !== 'darwin') { // TODO System tray on windows (maybe linux)
-    // since window-all-closed doesn't fire with our hidden audio process
-    player.win.once('closed', () => {
-      app.quit()
-    })
-  }
-
+  // register autoUpdater
   if (!process.env.DEV_SERVER) {
     setTimeout(() => {
       autoUpdater.checkForUpdatesAndNotify()
@@ -278,6 +130,149 @@ app.on('ready', function appReady () {
     log.info(`autoUpdater: Update downloaded`)
     broadcast('au:update-downloaded', info)
   })
+
+  // ACTIONS
+  // NOTE: I really don't like having all of these actions stuck in this scope.
+  // Would be nice to move this to a separate file eventually. -ungoldman
+
+  // Emit things to all windows
+  function broadcast (/* args */) {
+    var args = [].slice.call(arguments, 0)
+    windows.forEach((winObj) => {
+      if (winObj.win) winObj.win.send.apply(winObj.win, args)
+    })
+  }
+
+  function volume (ev, level) {
+    // player -> audio
+    state.volume = level
+    if (audio.win) audio.win.send('volume', level)
+  }
+
+  function queue (ev, newIndex) {
+    var newTrack = al.queue(newIndex)
+    log.info(newTrack)
+    broadcast('new-track', newTrack)
+    if (player.win) {
+      player.win.send('new-index', al.index)
+      player.win.send('is-new-query', al.isNewQuery)
+    }
+    if (state.playing) { broadcast('play') }
+    updateArtwork()
+  }
+
+  function updateArtwork () {
+    if (!get(al, `currentTrack.artwork`)) {
+      artwork.cache.getPath(al.currentKey, handleGetPath)
+    }
+  }
+
+  function handleGetPath (err, blobPath) {
+    if (err) return log.error(err)
+    al.currentTrack.artwork = blobPath
+    if (player.win) {
+      player.win.send('new-track', al.currentTrack)
+      player.win.send('new-index', al.index)
+    }
+  }
+
+  function play () {
+    state.playing = true
+    broadcast('play')
+  }
+
+  function pause () {
+    state.playing = false
+    broadcast('pause')
+  }
+
+  function playPause () {
+    state.playing ? pause() : play()
+  }
+
+  function prev () {
+    broadcast('new-track', al.prev())
+    if (player.win) player.win.send('new-index', al.index)
+    if (state.playing) { broadcast('play') }
+    updateArtwork()
+  }
+
+  function next () {
+    broadcast('new-track', al.next())
+    if (player.win) player.win.send('new-index', al.index)
+    if (state.playing) { broadcast('play') }
+    updateArtwork()
+  }
+
+  function mute (ev) {
+    state.muted = true
+    broadcast('mute')
+  }
+
+  function unmute (ev) {
+    state.muted = false
+    broadcast('unmute')
+  }
+
+  function timeupdate (ev, currentTime) {
+    // audio -> player
+    state.currentTime = currentTime
+    if (player.win) player.win.send('timeupdate', currentTime)
+  }
+
+  function shuffle (ev) {
+    al.shuffle()
+    if (player.win) player.win.send('shuffle')
+  }
+
+  function unshuffle (ev) {
+    al.unshuffle()
+    if (player.win) player.win.send('unshuffle')
+  }
+
+  function seek (ev, newTime) {
+    // player -> audio
+    state.currentTime = newTime
+    if (audio.win) audio.win.send('seek', newTime)
+  }
+
+  function handleNewTracks (err, newTrackDict) {
+    state.loading = false
+    broadcast('loading', false)
+    if (err) return log.warn(err)
+    var newState = al.load(newTrackDict)
+    if (player.win) player.win.send('track-dict', newState.trackDict, newState.order, state.paths)
+    console.timeEnd('update-library')
+    log.info('Done scanning. Found ' + Object.keys(newState.trackDict).length + ' tracks.')
+  }
+
+  function updateLibrary (ev, paths) {
+    if (state.loading) state.loading.destroy()
+    log.info('Updating library with new path(s): ' + paths)
+    console.time('update-library')
+    state.paths = paths
+    state.loading = makeTrackDict(paths, handleNewTracks)
+    broadcast('loading', true)
+  }
+
+  function search (ev, searchString) {
+    if (player.win) {
+      player.win.send('track-order', al.search(searchString))
+      player.win.send('is-new-query', al.isNewQuery)
+    }
+  }
+
+  function recall () {
+    if (player.win) player.win.send('recall', al.recall(), al.searchTerm)
+  }
+
+  function syncState (event) {
+    event.sender.send('sync-state', {
+      trackDict: al.trackDict,
+      order: al.order,
+      paths: state.paths
+    })
+  }
 })
 
 app.on('activate', function activate () {
@@ -307,8 +302,4 @@ app.on('before-quit', function beforeQuit (e) {
 
   libraryPersist.set(al.persist())
   app.quit()
-})
-
-process.on('uncaughtException', (err) => {
-  log.info(err)
 })
